report panel moved to its own folder
